+++ genson/__init__.py
@@ -1,2 +1,3 @@
-from .generator import Schema
-__all__ = ['Schema']
+from .schema.root import SchemaRoot
+from .schema.node import SchemaNode, InvalidSchemaError
+__all__ = ['SchemaRoot', 'SchemaNode', 'InvalidSchemaError']