+++ genson/schema/generators/__init__.py
@@ -1,0 +1,21 @@
+from .scalar import (
+    Typeless,
+    Null,
+    Boolean,
+    Number,
+    String
+)
+from .array import List, Tuple
+from .object import Object
+
+GENERATORS = (
+    Null,
+    Boolean,
+    Number,
+    String,
+    List,
+    Tuple,
+    Object
+)
+
+__all__ = tuple(list(GENERATORS) + [Typeless, GENERATORS])