+++ test/test_add_single.py
@@ -1,12 +1,7 @@
 from . import base
 
 
-class TestType(base.SchemaTestCase):
-
-    def test_no_schema(self):
-        schema = {}
-        self.add_schema(schema)
-        self.assertResult(schema)
+class TestType(base.SchemaNodeTestCase):
 
     def test_single_type(self):
         schema = {'type': 'string'}
@@ -18,43 +13,74 @@ class TestType(base.SchemaTestCase):
         self.add_schema(schema)
         self.assertResult(schema)
 
+    def test_typeless(self):
+        schema = {}
+        self.add_schema(schema)
+        self.assertResult(schema)
+
+    def test_array_type_no_items(self):
+        schema = {'type': 'array'}
+        self.add_schema(schema)
+        self.assertResult(schema)
+
+
+class TestAnyOf(base.SchemaNodeTestCase):
+
     def test_multi_type(self):
         schema = {'type': ['boolean', 'null', 'number', 'string']}
         self.add_schema(schema)
         self.assertResult(schema)
 
+    def test_multi_type_with_extra_keywords(self):
+        schema = {'type': ['boolean', 'null', 'number', 'string'],
+                  'title': 'this will be duplicated'}
+        self.add_schema(schema)
+        self.assertResult({'anyOf': [
+            {'type': 'boolean', 'title': 'this will be duplicated'},
+            {'type': 'null', 'title': 'this will be duplicated'},
+            {'type': 'number', 'title': 'this will be duplicated'},
+            {'type': 'string', 'title': 'this will be duplicated'}
+        ]})
 
-class TestPreserveKeys(base.SchemaTestCase):
-
-    def test_preserves_existing_keys(self):
-        schema = {'type': 'number', 'value': 5}
+    def test_anyof(self):
+        schema = {"anyOf": [
+            {"type": "null"},
+            {"type": "boolean", "title": "Gruyere"}
+        ]}
         self.add_schema(schema)
         self.assertResult(schema)
 
 
-class TestList(base.SchemaTestCase):
+class TestPreserveExtraKeywords(base.SchemaNodeTestCase):
 
-    def setUp(self):
-        base.SchemaTestCase.setUp(self)
-        self.set_schema_options(merge_arrays=True)
-
-    def test_add_items(self):
-        schema = {
-            'type': 'array',
-            'items': {'type': 'string'}}
+    def test_basic_type(self):
+        schema = {'type': 'boolean', 'const': False, 'myKeyword': True}
         self.add_schema(schema)
         self.assertResult(schema)
 
+    def test_number(self):
+        schema = {'type': 'number', 'const': 5, 'myKeyword': True}
+        self.add_schema(schema)
+        self.assertResult(schema)
 
-class TestTuple(base.SchemaTestCase):
+    def test_list(self):
+        schema = {'type': 'array', 'items': {"type": "null"},
+                  'const': [], 'myKeyword': True}
+        self.add_schema(schema)
+        self.assertResult(schema)
 
-    def setUp(self):
-        base.SchemaTestCase.setUp(self)
-        self.set_schema_options(merge_arrays=False)
+    def test_tuple(self):
+        schema = {'type': 'array', 'items': [{"type": "null"}],
+                  'const': [], 'myKeyword': True}
+        self.add_schema(schema)
+        self.assertResult(schema)
 
-    def test_add_items(self):
-        schema = {
-            'type': 'array',
-            'items': [{'type': 'string'}, {'type': 'null'}]}
+    def test_object(self):
+        schema = {'type': 'object', 'const': {}, 'myKeyword': True}
+        self.add_schema(schema)
+        self.assertResult(schema)
+
+    def test_typeless(self):
+        schema = {'const': 5, 'myKeyword': True}
         self.add_schema(schema)
         self.assertResult(schema)