+++ test/test_gen_multi.py
@@ -1,38 +1,56 @@
 from . import base
 
 
-class TestBasicTypes(base.SchemaTestCase):
+class TestBasicTypes(base.SchemaNodeTestCase):
 
     def test_single_type(self):
-        self.add_object('bacon')
-        self.add_object('egg')
-        self.add_object('spam')
-        self.assertResult({'type': 'string'})
-
-    def test_multi_type(self):
-        self.add_object('string')
-        self.add_object(1.1)
-        self.add_object(True)
-        self.add_object(None)
-        self.assertResult({'type': ['boolean', 'null', 'number', 'string']})
+        self.add_object("bacon")
+        self.add_object("egg")
+        self.add_object("spam")
+        self.assertResult({"type": "string"})
 
     def test_redundant_integer_type(self):
         self.add_object(1)
         self.add_object(1.1)
-        self.assertResult({'type': 'number'})
+        self.assertResult({"type": "number"})
 
 
-class TestArrayMerge(base.SchemaTestCase):
+class TestAnyOf(base.SchemaNodeTestCase):
+
+    def test_simple(self):
+        self.add_object("string")
+        self.add_object(1.1)
+        self.add_object(True)
+        self.add_object(None)
+        self.assertResult({"type": ["boolean", "null", "number", "string"]})
+
+    def test_complex(self):
+        self.add_object({})
+        self.add_object([None])
+        self.assertResult({"anyOf": [
+            {"type": "object"},
+            {"type": "array", "items": {"type": "null"}}
+        ]})
+
+    def test_simple_and_complex(self):
+        self.add_object(None)
+        self.add_object([None])
+        self.assertResult({"anyOf": [
+            {"type": "null"},
+            {"type": "array", "items": {"type": "null"}}
+        ]})
+
+
+class TestArrayList(base.SchemaNodeTestCase):
 
     def setUp(self):
-        base.SchemaTestCase.setUp(self)
-        self.set_schema_options(merge_arrays=True)
+        base.SchemaNodeTestCase.setUp(self)
 
     def test_empty(self):
         self.add_object([])
         self.add_object([])
 
-        self.assertResult({"type": "array", "items": {}})
+        self.assertResult({"type": "array"})
 
     def test_monotype(self):
         self.add_object(["spam", "spam", "spam", "eggs", "spam"])
@@ -69,17 +87,19 @@ class TestArrayMerge(base.SchemaTestCase):
         })
 
 
-class TestArrayPositional(base.SchemaTestCase):
-
-    def setUp(self):
-        base.SchemaTestCase.setUp(self)
-        self.set_schema_options(merge_arrays=False)
+class TestArrayTuple(base.SchemaNodeTestCase):
 
     def test_empty(self):
+        self.add_schema({"type": "array", "items": []})
+
         self.add_object([])
-        self.assertResult({"type": "array"})
+        self.add_object([])
+
+        self.assertResult({"type": "array", "items": [{}]})
 
     def test_multitype(self):
+        self.add_schema({"type": "array", "items": []})
+
         self.add_object([1, "2", "3", None, False])
         self.add_object([1, 2, "3", False])
 
@@ -95,6 +115,9 @@ class TestArrayPositional(base.SchemaTestCase):
         })
 
     def test_nested(self):
+        self.add_schema(
+            {"type": "array", "items": {"type": "array", "items": []}})
+
         self.add_object([
             ["surprise"],
             ["fear", "surprise"]
@@ -107,23 +130,13 @@ class TestArrayPositional(base.SchemaTestCase):
 
         self.assertResult({
             "type": "array",
-            "items": [
-                {
-                    "type": "array",
-                    "items": [
-                        {"type": "string"},
-                        {"type": "string"},
-                        {"type": "string"}
-                    ]
-                },
-                {
-                    "type": "array",
-                    "items": [
-                        {"type": "string"},
-                        {"type": "string"},
-                        {"type": "string"},
-                        {"type": "string"}
-                    ]
-                },
-            ]
+            "items": {
+                "type": "array",
+                "items": [
+                    {"type": "string"},
+                    {"type": "string"},
+                    {"type": "string"},
+                    {"type": "string"}
+                ]
+            }
         })